Changed JsonSchema to an interface
Fixes https://github.com/fge/json-schema-validator/issues/159
