+++ src/main/java/com/github/fge/jsonschema/main/JsonSchema.java
@@ -26,69 +26,15 @@ import com.github.fge.jsonschema.core.processing.Processor;
 import com.github.fge.jsonschema.core.report.ListProcessingReport;
 import com.github.fge.jsonschema.core.report.MessageProvider;
 import com.github.fge.jsonschema.core.report.ProcessingReport;
-import com.github.fge.jsonschema.core.report.ReportProvider;
-import com.github.fge.jsonschema.core.tree.SchemaTree;
-import com.github.fge.jsonschema.core.tree.SimpleJsonTree;
-import com.github.fge.jsonschema.processors.data.FullData;
-import com.github.fge.jsonschema.processors.validation.ValidationProcessor;
-
-import javax.annotation.concurrent.Immutable;
 
 /**
  * Single-schema instance validator
  *
- * <p>This is the class you will use the most often. It is, in essence, a {@link
+ * <p>This is the interface you will use the most often. It is, in essence, a {@link
  * JsonValidator} initialized with a single JSON Schema. Note however that this
  * class still retains the ability to resolve JSON References.</p>
- *
- * <p>It has no public constructors: you should use the appropriate methods in
- * {@link JsonSchemaFactory} to obtain an instance of this class.</p>
  */
-@Immutable
-public final class JsonSchema
-{
-    private final ValidationProcessor processor;
-    private final SchemaTree schema;
-    private final ReportProvider reportProvider;
-
-    /**
-     * Package private constructor
-     *
-     * @param processor the validation processor
-     * @param schema the schema to bind to this instance
-     * @param reportProvider the report provider
-     */
-    JsonSchema(final ValidationProcessor processor, final SchemaTree schema,
-        final ReportProvider reportProvider)
-    {
-        this.processor = processor;
-        this.schema = schema;
-        this.reportProvider = reportProvider;
-    }
-
-    private ProcessingReport doValidate(final JsonNode node,
-        final boolean deepCheck)
-        throws ProcessingException
-    {
-        final FullData data = new FullData(schema, new SimpleJsonTree(node),
-            deepCheck);
-        final ProcessingReport report = reportProvider.newReport();
-        final ProcessingResult<FullData> result
-            =  ProcessingResult.of(processor, report, data);
-        return result.getReport();
-    }
-
-    private ProcessingReport doValidateUnchecked(final JsonNode node,
-        final boolean deepCheck)
-    {
-        final FullData data = new FullData(schema, new SimpleJsonTree(node),
-            deepCheck);
-        final ProcessingReport report = reportProvider.newReport();
-        final ProcessingResult<FullData> result
-            =  ProcessingResult.uncheckedResult(processor, report, data);
-        return result.getReport();
-    }
-
+public interface JsonSchema {
     /**
      * Validate an instance and return a processing report
      *
@@ -102,12 +48,8 @@ public final class JsonSchema
      *
      * @since 2.1.8
      */
-    public ProcessingReport validate(final JsonNode instance,
-        final boolean deepCheck)
-        throws ProcessingException
-    {
-        return doValidate(instance, deepCheck);
-    }
+    ProcessingReport validate(JsonNode instance, boolean deepCheck)
+            throws ProcessingException;
 
     /**
      * Validate an instance and return a processing report
@@ -119,11 +61,8 @@ public final class JsonSchema
      * @return a processing report
      * @throws ProcessingException a processing error occurred during validation
      */
-    public ProcessingReport validate(final JsonNode instance)
-        throws ProcessingException
-    {
-        return validate(instance, false);
-    }
+    ProcessingReport validate(JsonNode instance)
+            throws ProcessingException;
 
     /**
      * Validate an instance and return a processing report (unchecked version)
@@ -148,12 +87,8 @@ public final class JsonSchema
      *
      * @since 2.1.8
      */
-    public ProcessingReport validateUnchecked(final JsonNode instance,
-        final boolean deepCheck)
-    {
-        return doValidateUnchecked(instance, deepCheck);
-    }
-
+    ProcessingReport validateUnchecked(JsonNode instance,
+                                              boolean deepCheck);
     /**
      * Validate an instance and return a processing report (unchecked version)
      *
@@ -164,10 +99,7 @@ public final class JsonSchema
      * @return a report (a {@link ListProcessingReport} if an exception was
      * thrown during processing)
      */
-    public ProcessingReport validateUnchecked(final JsonNode instance)
-    {
-        return doValidateUnchecked(instance, false);
-    }
+    ProcessingReport validateUnchecked(JsonNode instance);
 
     /**
      * Check whether an instance is valid against this schema
@@ -176,11 +108,8 @@ public final class JsonSchema
      * @return true if the instance is valid
      * @throws ProcessingException an error occurred during processing
      */
-    public boolean validInstance(final JsonNode instance)
-        throws ProcessingException
-    {
-        return doValidate(instance, false).isSuccess();
-    }
+    boolean validInstance(JsonNode instance)
+            throws ProcessingException;
 
     /**
      * Check whether an instance is valid against this schema (unchecked
@@ -192,8 +121,5 @@ public final class JsonSchema
      * @param instance the instance to validate
      * @return true if the instance is valid
      */
-    public boolean validInstanceUnchecked(final JsonNode instance)
-    {
-        return doValidateUnchecked(instance, false).isSuccess();
-    }
+    boolean validInstanceUnchecked(JsonNode instance);
 }
